requirements.txt
aiohappyeyeballs==2.4.0
aiohttp==3.10.5
aiohttp-retry==2.8.3
aiosignal==1.3.1
amqp==5.2.0
asgiref==3.8.1
attrs==24.2.0
billiard==4.2.0
celery==5.4.0
certifi==2024.7.4
charset-normalizer==3.3.2
click==8.1.7
click-didyoumean==0.3.1
click-plugins==1.1.1
click-repl==0.3.0
colorama==0.4.6
cron-descriptor==1.4.5
diff-match-patch==20230430
Django==5.1
django-background-tasks==1.2.8
django-celery-beat==2.7.0
django-import-export==4.1.1
django-timezone-field==7.0
frozenlist==1.4.1
idna==3.8
kombu==5.4.0
multidict==6.0.5
prompt_toolkit==3.0.47
psycopg2-binary==2.9.9
PyJWT==2.9.0
python-crontab==3.2.0
python-dateutil==2.9.0.post0
requests==2.32.3
six==1.16.0
sqlparse==0.5.1
tablib==3.5.0
twilio==9.2.4
tzdata==2024.1
urllib3==2.2.2
vine==5.1.0
wcwidth==0.2.13
yarl==1.9.4
